my_site_blog/requirements.txt
asgiref==3.8.1
Django==4.1
django-taggit==3.0.0
Markdown==3.4.1
pillow==11.0.0
psycopg2-binary==2.9.3
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2
